FS-7051: Preserve the annexb=no/yes status
As per RFC4856, Annex B mode for G.729 is disabled if the SDP contains a
fmtp parameter annexb=no and enabled if annexb=yes or the annexb parameter
is missing.

When responding to an offer mod_sangoma_codec always returns an answer
without a fmtp parameter. Even in response to a SDP offer which contains
one. This has the effect of responding to a request to disable Annex B
with an answer that only Annex B is supported.

With this change mod_sangoma_codec copies the fmtp parameter from the
offer to the answer fixing the problem.